Add merge support to git-rebase-branch.
Branch point logic has been simplified; we now find the first branch commit that is in the upstream log, and the first commit in the upstream reflog that is in the branch's log, and take whichever is most recent.

e.g. Suppose R1 - R2 - R3 - R4 <- BRANCH
                   \
                    R5 - R6

If UPSTREAM's reflog is (..., R3, R5, R6), the branch point is R3. (We assume R3 was amended.)
If UPSTREAM's reflog is (..., R3, R1, R5, R6), the branch point is R2. (We assume UPSTREAM was reset to R1.)