Fix bug causing unnecessary refreshes
Watchdog now fires events when files are opened or closed; filter these out, so we do not refresh when git metadata is read (this was causing an infinite loop). Additionally, fix an existing typo where directory events were supposed to be ignored but were not, causing git-watch-status to trigger its own refresh by creating a git lockfile.